V26_organizations_grading_method

+++ education-web/src/main/java/com/education/web/homework/dto/SubjectHomeworkProgressRow.java
@@ -1,12 +1,12 @@
 package com.education.web.homework.dto;
 
 /**
- * Агрегат по предмету: скільки ДЗ здано, скільки з них уже оцінено, сума зірок з оцінених.
+ * Агрегат по предмету: скільки ДЗ здано, скільки з них уже оцінено, показник з оцінених.
  */
 public record SubjectHomeworkProgressRow(
         String subject,
         int submittedCount,
         int gradedCount,
-        int starsTotal
+        double starsTotal
 ) {
 }
