+++ education-web/src/main/java/com/education/web/teacher/dto/TeacherGroupStudentStatRow.java
@@ -2,10 +2,10 @@ package com.education.web.teacher.dto;
 
 import java.util.Map;
 
-/** Один учень у групі: зірки за предметами (ключ — назва з teacher_subjects). */
+/** Один учень у групі: показник за предметами (sum або average — залежить від школи). */
 public record TeacherGroupStudentStatRow(
         String studentId,
         String fullName,
-        Map<String, Integer> starsBySubject
+        Map<String, Double> starsBySubject
 ) {
 }
